introduced device state view inheritance to decrease changing areas
Changelog: None
